+++ src/vector.odin
@@ -3,13 +3,13 @@ package main
 import "core:math"
 import "core:math/rand"
 
-Vec3 :: [3]f32
+Vec3 :: [3]f64
 
-vec3_length :: proc(v: Vec3) -> f32 {
+vec3_length :: proc(v: Vec3) -> f64 {
 	return math.sqrt(vec3_length_squared(v))
 }
 
-vec3_length_squared :: proc(v: Vec3) -> f32 {
+vec3_length_squared :: proc(v: Vec3) -> f64 {
 	return v.x * v.x + v.y * v.y + v.z * v.z
 }
 
@@ -22,7 +22,7 @@ vec3_normalize :: proc(v: Vec3) -> Vec3 {
 	return v / length
 }
 
-vec3_dot :: proc(a, b: Vec3) -> f32 {
+vec3_dot :: proc(a, b: Vec3) -> f64 {
 	return a.x * b.x + a.y * b.y + a.z * b.z
 }
 
@@ -30,7 +30,7 @@ vec3_cross :: proc(a, b: Vec3) -> Vec3 {
 	return Vec3{a.y * b.z - a.z * b.y, a.z * b.x - a.x * b.z, a.x * b.y - a.y * b.x}
 }
 
-vec3_distance :: proc(a, b: Vec3) -> f32 {
+vec3_distance :: proc(a, b: Vec3) -> f64 {
 	return vec3_length(b - a)
 }
